requirements.txt
grpcio==1.30.0
grpcio-tools==1.30.0
aiortc==1.0.0
python-socketio==4.6.0